+++ tests/src/NLP/Similarity/VectorSimTests.hs
@@ -1,12 +1,14 @@
 {-# LANGUAGE OverloadedStrings #-}
 module NLP.Similarity.VectorSimTests where
 
+import Prelude hiding (negate, sum)
 import Test.QuickCheck ( Property, (==>) )
 import Test.QuickCheck.Property ()
 import Test.Tasty (TestTree, testGroup)
 import Test.Tasty.QuickCheck (testProperty)
 
 import qualified Data.Text as T
+import qualified NLP.Similarity.VectorSim as TV
 
 import NLP.Similarity.VectorSim
 import NLP.Types (mkCorpus)
@@ -32,22 +34,22 @@ tests = testGroup "Vector Sim"
         --   , ("single", [1], 1)
         --   ]
          testGroup "tf tests" $ map (genTest2 tf)
-          [ ("", "test"::String , ["test"]::[String], 1)
-          , ("", "a", ["a", "test"], 1)
-          , ("", "a", ["test"], 0)
+          [ ("", "test", mkDocument ["test"], 1)
+          , ("", "a",    mkDocument ["a", "test"], 1)
+          , ("", "a",    mkDocument ["test"], 0)
           ]
         , testGroup "idf tests" $ map (genTestF2 idf)
-          [ ("", "test", mkCorpus [["test"]], log(1/2))
+          [ ("", "test", mkCorpus [["test"]], log(2/3))
           , ("", "a", mkCorpus [["test"]], log(1))
           , ("", "a", mkCorpus [["a", "test"],["test"]], log(2/2))
           ]
         , testGroup "tf_idf tests" $ map (genTestF3 tf_idf)
-          [ ("", "test", ["test"], mkCorpus [["test"]], log(2/3))
-          , ("", "a", ["a", "test"], mkCorpus [["some"], ["test"]], 0.40546)
-          , ("", "foo", ["foo"], mkCorpus [["test"]], 0)
-          , ("", "foo", ["foo"], mkCorpus [["foo"],["test"]], 0)
-          , ("", "bar", ["foo"], mkCorpus [["test"]], 0)
-          , ("", "bar", ["foo"], mkCorpus [["foo"],["test"]], 0)
+          [ ("", "test", mkDocument ["test"], mkCorpus [["test"]], log(2/3))
+          , ("", "a",    mkDocument ["a", "test"], mkCorpus [["some"], ["test"]], 0.40546)
+          , ("", "foo",  mkDocument ["foo"], mkCorpus [["test"]], 0)
+          , ("", "foo",  mkDocument ["foo"], mkCorpus [["foo"],["test"]], 0)
+          , ("", "bar",  mkDocument ["foo"], mkCorpus [["test"]], 0)
+          , ("", "bar",  mkDocument ["foo"], mkCorpus [["foo"],["test"]], 0)
           ]
         , testGroup "Similarity tests, trivial corpus" $
             map (genTest2 $ sim $ mkCorpus [["test"]])
@@ -71,7 +73,24 @@ tests = testGroup "Vector Sim"
         , testProperty "idf /= NaN" prop_idfIsANum
         , testProperty "tf_idf /= NaN" prop_tf_idfIsANum
         , testProperty "similarity /= NaN" prop_similarity_isANum
+        , testProperty "v + 0 = v" prop_addVectorZero
+        , testProperty "v - v = 0" prop_negateVector
         ]
+
+prop_addVectorZero :: TermVector -> Bool
+prop_addVectorZero v = addVectors v zeroVector == v &&
+                       addVectors zeroVector v == v
+
+prop_negateVector :: TermVector -> Bool
+prop_negateVector v =
+    let theSum = addVectors v (negate v)
+    in and [ TV.lookup k theSum == 0
+           | k <- TV.keys v
+           ]
+
+prop_sum :: [TermVector] -> Bool
+prop_sum vs =
+    foldr addVectors zeroVector vs == sum vs
 
 prop_idfIsANum :: String -> [[String]] -> Bool
 prop_idfIsANum term docs = not (isNaN (idf termTxt $ mkCorpus docsTxt))
@@ -80,7 +99,7 @@ prop_idfIsANum term docs = not (isNaN (idf termTxt $ mkCorpus docsTxt))
     docsTxt = map (map T.pack) docs
 
 prop_tf_idfIsANum :: String -> [String] -> [[String]] -> Bool
-prop_tf_idfIsANum term doc docs = not $ isNaN $ tf_idf termTxt docTxt $ mkCorpus docsTxt
+prop_tf_idfIsANum term doc docs = not $ isNaN $ tf_idf termTxt (mkDocument docTxt) $ mkCorpus docsTxt
   where
     termTxt = T.pack term
     docTxt = map T.pack doc