+++ changelog.md
@@ -1,3 +1,40 @@
+= HEAD =
+ - Added deriving clauses for Class and Feature to support GHC 7.10;
+   based on Rick Dzekman's pull request.
+
+= 0.8.0.0 =
+
+ - Added a Document type to better store term count indexes
+   (essentially TermVectors with integer values).
+
+ - Used the Document type for tf_idf; this changed the API, and is the
+   reason for the B-level version bump. (also done for performance)
+
+ - A number of non-breaking performance improvements (e.g., not using foldl in a few places.)
+
+ - These changes, and much of 0.7.0.0 are due to work by Tristan Ravitch & Jonathan Daugherty.
+
+= 0.7.0.0 =
+ - B-level version bump because we added test dependency on
+   unordered-containers, which could cause downstream issues.
+
+ - TermVector: is a newtype now, and has its own arbitrary instance.
+
+ - TermVector: adds addVectors, zeroVector, negate, and sum
+
+ - DefaultMap: adds elems, map, and unionWith
+
+ - Adds quickcheck properties for all of the above
+
+= 0.6.0.0 =
+
+ - Switched to using Hashmap for the DefaultMap implementation;
+   this *may* break compatibility with old binary stores.
+
+= 0.5.2.1 =
+
+ - Removed an orphan instance for Text.
+
 = 0.5.2.0 =
 
  - Unceremoniously forced the Chunking model to do basic Named